Trade protocol is now completely implemented.
Orders now progress through the various stages at the correct times.
Order manifest fetching and processing is now done in parallel.
Fixed a load of logic bugs in the account and order managers.
Fixed a bug where orders with the same item would overwrite each other
in the order items table. Resolved by using unique IDs for each row
instead of using the ItemCode.
Fixed a bug where delivered items just vanished into the void.
Fixed a bug where Order Items and Orders for each colony where not
cleared when the save game was loaded.
Fixed a deadlock when updating the order status in many parallel tasks.
Order list is completely re-downloaded when ever the save is loaded.
Order manifests are now only ever downloaded when the save is loaded or
when the order is placed.
Added lots of new error handling and trapping to ensure the managers
keep retrying and not just give up immediately.